\name{mfaces-internal}
\alias{kr}
\alias{pspline.setting}
\alias{raw.construct}
\alias{construct.knots}
\alias{check.data}
\alias{matrix.multiply}
\alias{bps}
\alias{igcv.wrapper}
\alias{igcv.criteria}
\alias{face.sparse.inner}
\alias{predict.face.sparse.inner}
\title{Internal functions for mfaces package}
\description{
Internal function.
}
