\name{mfaces-package}
\alias{mfaces-package}
\alias{mfaces}
\docType{package}
\title{mfaces}
\description{
Fast Covariance Estimation for Multivariate Sparse Functional Data
}
\details{
\tabular{ll}{
Package: \tab mfaces\cr
Type: \tab Package\cr
Version: \tab 0.1-4\cr
Date: \tab 2022-07-18\cr
License: \tab GPL-3
}
}
\author{
Cai Li and Luo Xiao

Maintainer: Cai Li <cai.li.stats@gmail.com>
}
\references{
Cai Li, Luo Xiao, and Sheng Luo, 2020. Fast covariance estimation for multivariate sparse functional data. Stat, 9(1), p.e245, 
\doi{10.1002/sta4.245}.
}

\keyword{ package }

